requirements.txt
Flask>=2.3.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0
scikit-image>=0.22.0 